third_party/WebKit/LayoutTests/media/autoplay/console-warning-gesture-media-expected.txt

CONSOLE WARNING: line 2: <iframe gesture="media"> is not supported. Use <iframe allow="autoplay">, https://goo.gl/ximf56
